backend/requirements.txt
aiosmtplib==3.0.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.2.1
certifi==2025.1.31
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.0
email-validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
mako==1.3.9
markupsafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1-modules==0.4.1
pydantic==2.8.0
pydantic-core==2.20.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
rsa==4.9
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.38
starlette==0.45.3
typing-extensions==4.12.2
uvicorn==0.34.0
